+++ CommonCrypto/Source/ccOpenssl/des_locl.h
@@ -172,7 +172,7 @@
 #  define ROTATE(a,n)	({ register unsigned int ret;	\
 				asm ("rorl %1,%0"	\
 					: "=r"(ret)	\
-					: "I"(n),"0"(a)	\
+					: "I"(n),"0"((unsigned int)a)	\
 					: "cc");	\
 			   ret;				\
 			})
