+++ apr/apr-util/apr-util/dbm/sdbm/sdbm_pair.c
@@ -1,9 +1,9 @@
-/* Copyright 2000-2005 The Apache Software Foundation or its licensors, as
- * applicable.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+/* Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -183,7 +183,7 @@ apr_sdbm_datum_t key;
 		register int m;
 		register char *dst = pag + (i == 1 ? PBLKSIZ : ino[i - 1]);
 		register char *src = pag + ino[i + 1];
-		register int   zoo = dst - src;
+		register short zoo = (short) (dst - src);
 
 		debug(("free-up %d ", zoo));
 /*
